add confirm popup

+++ front-end/src/components/productlist.js
@@ -1,7 +1,49 @@
 import React, { useState , useEffect } from "react";
 import { Link } from "react-router-dom";
+import {  Modal, Space } from 'antd';
+import $ from 'jquery';
+//add
+import { styled } from '@mui/material/styles';
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell, { tableCellClasses } from '@mui/material/TableCell';
+import TableContainer from '@mui/material/TableContainer';
+import TableHead from '@mui/material/TableHead';
+import TableRow from '@mui/material/TableRow';
+import Paper from '@mui/material/Paper';
+
+import Stack from '@mui/material/Stack';
+import Button from '@mui/material/Button'; 
+
+import TextField from '@mui/material/TextField';
+
+const { confirm } = Modal;
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.common.black,
+    color: theme.palette.common.white,
+  },
+  [`&.${tableCellClasses.body}`]: {
+    fontSize: 14,
+  },
+}));
+
+const StyledTableRow = styled(TableRow)(({ theme }) => ({
+  '&:nth-of-type(odd)': {
+    backgroundColor: theme.palette.action.hover,
+  },
+  // hide last border
+//   '&:last-child td, &:last-child th': {
+//     border: 0,
+//   },
+}));
+
+//end
+
 
 const ListProduct = ()=>{
+  $('.ant-fade').hide();
+  $('.ant-modal-root').hide();
     const [products,setProducts] = useState([]);
 
 // LOAD WHEN COMPONETS CALL
@@ -15,30 +57,38 @@ const ListProduct = ()=>{
               authorization:`bearer ${JSON.parse(localStorage.getItem('token'))}`
             },
         });
-        // no need to use method ,headers etc due to get mehtod
-        //    method : 'post',
-        //    body: JSON.stringify({nam,price,category,company}),
-        //    headers: {"Content-Type":"application/json"}
          result = await result.json();
          setProducts(result);
     }
     console.warn('produ-list:',products);
 
-    const deleteProject = async (id)=>{   
-         //console.warn(id);
-        let data = await fetch(`http://localhost:4000/delete/${id}`,{
+
+    const deletepro = async (id)=>{
+      $('.ant-modal-root').show();
+      $('.ant-fade').show();
+      
+      confirm({
+        title: 'Do you Want to delete this Product?',
+        //content: 'Some descriptions',
+        onOk:  async ()=>{
+          let data = await fetch(`http://localhost:4000/delete/${id}`,{
             method: 'delete',    
             headers: {
                 authorization:`bearer ${JSON.parse(localStorage.getItem('token'))}`
               }
-        });
-        data = await data.json();
-        if(data){
-            getProduct(); 
-        alert("Dlete data successful");
-      }else alert('no found')
-    }
-
+            });
+            data = await data.json();
+            if(data){
+                getProduct();
+          }else alert('no found')
+        },
+        onCancel() {
+          console.log('Cancel');
+          $('.ant-fade').hide();
+          $('.ant-modal-root').hide();
+        },
+      });
+  }
     const searchitem = async (ee)=>{
         //console.warn(ee.target.value)
         let itemValue = ee.target.value;
@@ -58,40 +108,46 @@ const ListProduct = ()=>{
     }
     return(
         <div className="product-list-container">
-            <input type="text" placeholder="Search item" onChange={searchitem}/>
-        <table>
-           <thead>
-           <tr>
-           <th>sr. Number</th>
-            <th>Product name</th>
-            <th>Price</th>
-            <th>Category</th>
-            <th>company</th>
-            <th>Actions</th>
-            </tr>
-           </thead>
-           <tbody>
-           {   
-            products.length > 0 ? products.map((items,index)=>
-             <>
-                <tr key={items}>
-
-                    <td>{index+1}</td>
-                    <td>{items.name}</td>
-                    <td>{items.price}</td>
-                    <td>{items.category}</td>
-                    <td>{items.company}</td>
-                    <td><button type="button" onClick={()=>deleteProject(items._id)}>Delete</button>
-                        <Link to={'/update/'+items._id}>update</Link>
-                    </td>
-                </tr>
-            </>   
-                            )
-             : <tr><td>No reuslt found</td></tr>
-
-                        }
-           </tbody>
-        </table>
+                   <TextField
+                        label="Search field"
+                        type="search"
+                        variant="standard"
+                        onChange={searchitem}
+        />   
+            <TableContainer component={Paper}>
+      <Table sx={{ minWidth: 700 }} aria-label="customized table">
+        <TableHead>
+          <TableRow>
+            <StyledTableCell align="center">sr. Number</StyledTableCell>
+            <StyledTableCell align="center">Product name</StyledTableCell>
+            <StyledTableCell align="center">Price</StyledTableCell>
+            <StyledTableCell align="center">Category</StyledTableCell>
+            <StyledTableCell align="center">Company</StyledTableCell>
+            <StyledTableCell align="center">Actions</StyledTableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {
+          products.length > 0 ? products.map((items,index)=> (
+            <StyledTableRow key={items.name}>
+              <StyledTableCell align="center">{index+1}</StyledTableCell>
+              <StyledTableCell align="center">{items.name}</StyledTableCell>
+              <StyledTableCell align="center">{items.price}</StyledTableCell>
+              <StyledTableCell align="center">{items.category}</StyledTableCell>
+              <StyledTableCell align="center">{items.company}</StyledTableCell>
+              <StyledTableCell align="center">
+                <Stack spacing={1} direction="row">
+                        <Button variant="contained" onClick={()=>deletepro(items._id)}>Delete</Button>
+                        <Button variant="outlined" color="success"><Link to={'/update/'+items._id}>update</Link></Button>
+                </Stack>
+              </StyledTableCell>
+            </StyledTableRow>
+          ) ):
+          <StyledTableRow><StyledTableCell align="center" colSpan={6}>no data found</StyledTableCell></StyledTableRow>
+          }
+        </TableBody>
+      </Table>
+    </TableContainer> 
         </div>
     )
 }
